Teensy 4.1 port with Ethernet support (#144)
* Start porting to Teensy 4.1, also fix some compiler warnings in FSL HAL

* Add hardware init for Teensy, refactor phy drivers, rework IMX EMAC autonegotiate code

* Revert some testing changes

* Fix incorrect phy address used outside of low_level_init_successful()

* Ethernet operational!

* Turn off DEBUG_IMX_EMAC

* Style fixes

* Style again

* Bugfix: mbed_lib.json files in project source dir were not getting picked up

* Bugfix: CLion debug configurations not generated properly due to variables going out of scope

* Support Teensy OTP MAC address

+++ targets/TARGET_NXP/TARGET_MCUXpresso_MCUS/TARGET_MIMXRT1050/drivers/fsl_dcp.c
@@ -96,7 +96,7 @@ static dcp_context_t s_dcpContextSwitchingBuffer;
 
 static void dcp_reverse_and_copy(uint8_t *src, uint8_t *dest, size_t src_len)
 {
-    for (int i = 0; i < src_len; i++)
+    for (size_t i = 0; i < src_len; i++)
     {
         dest[i] = src[src_len - 1 - i];
     }
