+++ targets/TARGET_NXP/TARGET_MCUXpresso_MCUS/TARGET_MIMXRT1050/drivers/fsl_flexram.c
@@ -200,7 +200,7 @@ status_t FLEXRAM_AllocateRam(flexram_allocate_ram_t *config)
             continue;
         }
 
-        if (i < (dtcmBankNum + ocramBankNum + itcmBankNum))
+        if (i < (uint32_t)(dtcmBankNum + ocramBankNum + itcmBankNum))
         {
             bankCfg |= ((uint32_t)kFLEXRAM_BankITCM) << (i * 2);
             continue;
